Update index.html
Modified some links from target="_self" to target="_blank", so that when the user clicks on the link of a repo or other website, a new window will open. This way the user is directed to the link and he/she stays on the website also. Also added some minor changes.